api-admin/requirements.txt
flask
flask-cors
mysql-connector-python
bcrypt
PyJWT
